fix decompression bug in fastboot
fastboot passes the *uncompressed* length of the file as the length of
the input to the inflate() call, which happens to work unless the
compressed data is actually larger than the uncompressed data (which
it can be for very small files).  Fix this to pass the correct
compressed length down to the inflate call.