Better support 64bit x86 kernel crashes.
FindCrashingFunction() fails to find the crashing kernel function because
the x86 64bit kernel code (arch/x86/kernel/dumpstack.c:die()) issues "RIP"
instead of "EIP".  So it falls back to using the panic string in the crash
signature.

This causes the logging_KernelCrash.py test to fail because it looks for the
crashing function in the signature string.

BUG=chromium-os:22353
TEST=logging_KernelCrash

Change-Id: Ic4790648facf5f0be2b82001f8e18b606eca16c7
Reviewed-on: https://gerrit.chromium.org/gerrit/19492
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Bryan Freed <bfreed@chromium.org>
Commit-Ready: Bryan Freed <bfreed@chromium.org>