add -f option to mkbootfs to load file perms from file
mkbootfs uses the compiled in fs_config() function to set the mode
(but not the uid or gid) of the files in the filesystem it is
building.  Give it an optional "-f filename" flag to allow loading the
list of file permissions from a file, so we can continue to accurately
recreate old filesystems even when the current fs_config() has changed.

Bug: 6435132
Change-Id: Ife6b25e989301e78e22c34a544738436c96caf6f