Remove UniquePtr.h copy
This was copied from libcore/include quite a while ago, but the
canonical version has since moved out to a generic library called
libnativehelper. All users of this header should already have
libnativehelper on their include path, so switching to the canonical
version is as easy as removing the "utils/" part.

Change-Id: Iae8e59bf3eee573bfa78381866989934e5bbf19d