Dispose of ash.
We haven't used ash for several releases. Remove ash and its
companion liblinenoise.

I've left the support for an alternative shell here for now,
but we can move that out to external/mksh later if we want.
(For now this seemed like a somewhat useful signpost to the real shell.)

Change-Id: I634c7969f6cbe375a2bc56fc1077e64888380f3d